requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cycler==0.11.0
Cython==0.29.32
Deprecated==1.2.13
Django==4.1
django-debug-toolbar==3.6.0
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
download==0.3.5
flower==1.2.0
fonttools==4.37.1
humanize==4.3.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
kombu==5.2.4
libsvmdata==0.4.1
Markdown==3.4.1
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.0
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xarray==2022.6.0
